% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_indicators.R
\name{ortho}
\alias{ortho}
\title{Residual-centered product indicators}
\usage{
ortho(data, vars1, vars2, bind = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{bind}{Logical. If \code{TRUE}, the product indicators are
appended to the data frame provided by \code{data}. If
\code{FALSE}, only the product indicators are returned.}

\item{var1, var2}{Character vectors naming the indicators to be
multiplied together.}
}
\value{
A data frame. If \code{bind = TRUE}, the product indicators
are appended to the input data frame given by \code{data}. If
\code{FALSE}, only the product indicators are returned.
}
\description{
Creates residual-centered product indicators.
}

